Introduce Drum engine family and integrate md-drum-synth models into runtime/UI

+++ BRICK6_CUBE_fonctionnel/md-drum-synth-main/FmCowbellModel.cpp
@@ -20,7 +20,7 @@ static float WrapPhase(float phase) {
 }
 
 static float ExpDecay(float t, float decay_time) {
-    return std::expf(-t / decay_time);
+    return std::exp(-t / decay_time);
 }
 
 void FmCowbellModel::Init() {
