+++ BRICK6_CUBE_fonctionnel/md-drum-synth-main/FmRimshotModel.cpp
@@ -17,7 +17,7 @@ float WrapPhase(float phase) {
 }
 
 float ExpDecay(float t, float decay_time) {
-    return std::expf(-t / decay_time);
+    return std::exp(-t / decay_time);
 }
 
 void FmRimshotModel::Init() {
